(add): user crud test
